Make the monitor accessible from C#
