+++ ksp_plugin/ksp_plugin.vcxproj
@@ -370,6 +370,7 @@ copy /Y "$(OutDir)$(TargetName).pdb" "$(SolutionDir)$(Configuration)\GameData\Pr
     <ClCompile Include="interface_flight_plan.cpp" />
     <ClCompile Include="interface_future.cpp" />
     <ClCompile Include="interface_iterator.cpp" />
+    <ClCompile Include="interface_monitor.cpp" />
     <ClCompile Include="interface_planetarium.cpp" />
     <ClCompile Include="interface_renderer.cpp" />
     <ClCompile Include="interface_vessel.cpp" />
