+++ ksp_plugin_adapter/ksp_plugin_adapter.cs
@@ -1260,6 +1260,7 @@ public partial class PrincipiaPluginAdapter
             is_post_apocalyptic_ |=
                 plugin_.HasEncounteredApocalypse(out revelation_);
           }
+          Interface.MonitorStart();
           foreach (var vessel in FlightGlobals.Vessels) {
             if (vessel.packed && plugin_.HasVessel(vessel.id.ToString())) {
               vessel_futures_.Add(
@@ -1313,6 +1314,7 @@ public partial class PrincipiaPluginAdapter
         Interface.FutureWait(ref future);
       }
       vessel_futures_.Clear();
+      Interface.MonitorStop();
       ApplyToVesselsOnRails(
           vessel => UpdateVessel(vessel, Planetarium.GetUniversalTime()));
     }
